Refactor shared code into utils.ts
